requirements.txt
pytest
requests
colorama
python-dotenv